Fix typo (Blame IDEA) and move mixin
Moved the Mixin to The Goal Section Because its a Goal
