+++ src/main/java/carpetfixes/mixins/goalFixes/FoxEntity$JumpChasingGoal_TickMixin.java
@@ -1,4 +1,4 @@
-package carpetfixes.mixins.entityFixes;
+package carpetfixes.mixins.goalFixes;
 
 
 import carpetfixes.CFSettings;
@@ -13,11 +13,11 @@ import org.spongepowered.asm.mixin.injection.Redirect;
 @Mixin(targets = "net.minecraft.entity.passive.FoxEntity$JumpChasingGoal")
 public class FoxEntity$JumpChasingGoal_TickMixin {
 
-    
+
     @Redirect(method = "tick",at=@At(value = "INVOKE",target = "Lnet/minecraft/block/BlockState;isOf(Lnet/minecraft/block/Block;)Z"))
     public boolean FoxEntity_SnowCheckFix(BlockState instance, Block block) {
 
-            return instance.getBlock().equals(Blocks.SNOW) || (CFSettings.foxesDisregardPowderSnowFix && instance.getBlock().equals(Blocks.POWDER_SNOW));
+            return instance.getBlock().equals(Blocks.SNOW) || (CFSettings.foxesDisreguardPowderSnowFix && instance.getBlock().equals(Blocks.POWDER_SNOW));
 
     }
