+++ src/main/resources/carpet-fixes.mixins.json
@@ -182,7 +182,6 @@
     "entityFixes.FallingBlockEntity_netherPortalMixin",
     "entityFixes.FishingBobberEntity_outsideWaterMixin",
     "entityFixes.FoxEntity_lootDropMixin",
-    "entityFixes.FoxEntity$JumpChasingGoal_TickMixin",
     "entityFixes.HopperMinecartEntity_slowMixin",
     "entityFixes.ItemFrameEntity_explosionWaterMixin",
     "entityFixes.ItemFrameEntity_soundMixin",
@@ -254,6 +253,7 @@
     "goalFixes.EndermanEntity$PlaceBlockGoal_updatePlaceMixin",
     "goalFixes.FollowParentGoal_movementMixin",
     "goalFixes.FoxEntity$AvoidDaylightGoal_originMixin",
+    "goalFixes.FoxEntity$JumpChasingGoal_TickMixin",
     "goalFixes.MobEntity_stopGoalOnDeathMixin",
     "goalFixes.OcelotEntity$FleeGoal_vehicleMixin",
     "goalFixes.ProjectileAttackGoal_deadMixin",
